# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateMenuRequest:
    properties:
      name:
        type: string
      order:
        type: integer
      parent_id:
        type: string
    required:
    - name
    type: object
  dto.MenuResponse:
    properties:
      children:
        items:
          $ref: '#/definitions/dto.MenuResponse'
        type: array
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      order:
        type: integer
      parent_id:
        type: string
      updated_at:
        type: string
    type: object
  dto.MoveMenuRequest:
    properties:
      parent_id:
        type: string
    type: object
  dto.ReorderMenuRequest:
    properties:
      order:
        type: integer
    type: object
  dto.UpdateMenuRequest:
    properties:
      name:
        type: string
      order:
        type: integer
      parent_id:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Production-ready hierarchical menu management system.
  title: Menu Tree Management API
  version: "1.0"
paths:
  /api/menus:
    get:
      description: Get all menus organized in a hierarchical tree
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.MenuResponse'
                  type: array
              type: object
      summary: Get all menus in tree structure
      tags:
      - menus
    post:
      consumes:
      - application/json
      description: Create a new menu item
      parameters:
      - description: Menu object
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMenuRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MenuResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create a new menu
      tags:
      - menus
  /api/menus/{id}:
    delete:
      description: Delete a menu item and all its children recursively
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete a menu
      tags:
      - menus
    get:
      description: Get a menu item by ID
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MenuResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get a single menu
      tags:
      - menus
    put:
      consumes:
      - application/json
      description: Update a menu item's details
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      - description: Menu update object
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MenuResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update a menu
      tags:
      - menus
  /api/menus/{id}/move:
    patch:
      consumes:
      - application/json
      description: Move a menu item to a different parent
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      - description: Move request
        in: body
        name: move
        required: true
        schema:
          $ref: '#/definitions/dto.MoveMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Move a menu item
      tags:
      - menus
  /api/menus/{id}/reorder:
    patch:
      consumes:
      - application/json
      description: Change the display order of a menu item within its current level
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      - description: Reorder request
        in: body
        name: reorder
        required: true
        schema:
          $ref: '#/definitions/dto.ReorderMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Reorder a menu item
      tags:
      - menus
swagger: "2.0"